requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
apscheduler==3.10.4
python-multipart==0.0.12
playwright==1.47.0
